+++ src/gui/widgets/dialogs/gw_impl_ui.cpp
@@ -176,6 +176,7 @@ void GlaxnimateWindow::Private::init_actions()
     connect(ui.action_quit, &QAction::triggered, parent, &GlaxnimateWindow::close);
     connect(ui.action_move_to, &QAction::triggered, parent, &GlaxnimateWindow::move_to);
     connect(ui.action_validate_tgs, &QAction::triggered, parent, &GlaxnimateWindow::validate_tgs);
+    connect(ui.action_validate_discord, &QAction::triggered, parent, [this]{validate_discord();});
     connect(ui.action_resize, &QAction::triggered, parent, [this]{ ResizeDialog(this->parent).resize_document(current_document.get()); });
     connect(ui.action_donate, &QAction::triggered, parent, &GlaxnimateWindow::help_donate);
     connect(ui.action_new_layer, &QAction::triggered, parent, [this]{layer_new_layer();});