+++ src/gui/widgets/dialogs/io_status_dialog.cpp
@@ -22,6 +22,7 @@ IoStatusDialog::IoStatusDialog(const QIcon& icon, const QString& title, bool del
     d->icon = icon;
     d->setupUi(this);
     setWindowTitle(title);
+    setWindowIcon(icon);
 }
 
 IoStatusDialog::~IoStatusDialog() = default;